# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Axion API
  version: "1.0.0"
  description: |
    Axion — The Operating System for the Autonomous Enterprise.

    Hire AI employees, orchestrate them as workflows, audit every action.
  contact:
    name: Axion Engineering
    email: engineering@axion.ai
    url: https://axion.ai
  license:
    name: Proprietary
servers:
  - url: https://api.axion.ai/v1
    description: Production
  - url: https://api.staging.axion.ai/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Local
security:
  - bearerAuth: []
  - bearerAuth: []
    orgHeader: []
...
paths:
  /auth/signup:
    post:
      tags: [auth]
      summary: Create user + initial org, return token pair
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UserCreate" } } }
      responses:
        "201": { description: created, content: { application/json: { schema: { $ref: "#/components/schemas/TokenPair" } } } }
        "409": { description: email already in use }
...
  /agents/catalog:
    get:
      tags: [agents]
      summary: List agent templates
      parameters:
        - { in: query, name: category, schema: { type: string } }
        - { in: query, name: q, schema: { type: string } }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AgentTemplate" }
  /agents/catalog/{slug}:
    get:
      tags: [agents]
      parameters:
        - { in: path, name: slug, required: true, schema: { type: string } }
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/AgentTemplate" } } } }
        "404": { description: not found }
  /agents:
    get:
      tags: [agents]
      parameters:
        - { in: query, name: status, schema: { $ref: "#/components/schemas/AgentStatus" } }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AgentInstance" }
    post:
      tags: [agents]
      summary: Hire a new agent
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AgentInstanceCreate" } } }
      responses:
        "201": { description: created, content: { application/json: { schema: { $ref: "#/components/schemas/AgentInstance" } } } }
  /agents/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [agents]
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/AgentInstance" } } } }
    patch:
      tags: [agents]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AgentInstanceUpdate" } } }
      responses:
        "200": { description: ok, content: { application/json: { schema: { $ref: "#/components/schemas/AgentInstance" } } } }
    delete:
      tags: [agents]
      responses: { "204": { description: deleted } }
  /agents/{id}/run:
    post:
      tags: [agents]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [objective]
              properties:
                objective: { type: string }
                inputs: { type: object, additionalProperties: true }
                timeout_s: { type: integer, default: 300, minimum: 10, maximum: 3600 }
      responses:
        "202": { description: accepted, content: { application/json: { schema: { $ref: "#/components/schemas/Task" } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    orgHeader:
      type: apiKey
      in: header
      name: X-Org-Id
  schemas:
    UserCreate:
      type: object
      required: [email, full_name, password]
      properties:
        email: { type: string, format: email }
        full_name: { type: string }
        password: { type: string, format: password, minLength: 10 }
        org_name: { type: string }
...
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, default: bearer }
        expires_in: { type: integer }
...
    AgentStatus:
      type: string
      enum: [idle, running, paused, error, archived]
    AgentTemplate:
      type: object
      properties:
        slug: { type: string }
        name: { type: string }
        role: { type: string }
        category: { type: string }
        description: { type: string }
        icon: { type: string }
        skills: { type: array, items: { type: string } }
        default_tools: { type: array, items: { type: string } }
        price_eur_monthly: { type: integer }
    AgentInstanceCreate:
      type: object
      required: [template_slug, name]
      properties:
        template_slug: { type: string }
        name: { type: string }
        config: { type: object, additionalProperties: true }
        enabled_tools: { type: array, items: { type: string } }
        custom_prompt: { type: string }
        voice_clone_id: { type: string }
        budget_per_day_eur: { type: integer, default: 100 }
    AgentInstanceUpdate:
      type: object
      properties:
        name: { type: string }
        config: { type: object, additionalProperties: true }
        enabled_tools: { type: array, items: { type: string } }
        custom_prompt: { type: string }
        budget_per_day_eur: { type: integer }
        status: { $ref: "#/components/schemas/AgentStatus" }
    AgentInstance:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        template_slug: { type: string }
        name: { type: string }
        status: { $ref: "#/components/schemas/AgentStatus" }
        config: { type: object, additionalProperties: true }
        enabled_tools: { type: array, items: { type: string } }
        budget_per_day_eur: { type: integer }
        health_score: { type: number }
        tasks_today: { type: integer }
        created_at: { type: string, format: date-time }
    Task:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        agent_id: { type: string, format: uuid }
        objective: { type: string }
        status: { type: string, enum: [queued, running, succeeded, failed, cancelled] }
        input_payload: { type: object }
        output_payload: { type: object }
        duration_ms: { type: integer, nullable: true }
        tokens_in: { type: integer }
        tokens_out: { type: integer }
        cost_eur: { type: number }
        model_used: { type: string, nullable: true }
        error: { type: string, nullable: true }
        trace_id: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
...